Refactored code and reorganized files
